remove container for date time range picker
